Add in difficulty and level selection code
The proper level should now be loaded upon enter the level scene based on the difficulty and level static variables. Also, bullets will now point in the direction they are firing if checked in the enemyPattern objects. (Make sure t o import bullets that arent spherical pointing towards the right for the effect to work.)